refactor: 💡 delete providers
