+++ src/app/(dumper)/@contents/page.tsx
@@ -3,7 +3,7 @@ import { Separator } from "@/components/ui/separator";
 import { checkSelfAuthOrRedirectToAuth } from "@/features/auth/utils/get-session";
 import { hasDumperPermission } from "@/features/auth/utils/role";
 import { AddContentsProvider } from "@/features/contents/components/add-contents-provider";
-import { ContentsStackProvider } from "@/features/contents/components/contents-stack-provider";
+import { ContentsStack } from "@/features/contents/components/contents-stack";
 import { AddFormSkeleton } from "@/features/dump/components/add-form-skeleton";
 import { Suspense } from "react";
 
@@ -22,7 +22,7 @@ export default async function Page() {
 			)}
 			<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
 			<Suspense fallback={<CardStackSkeleton />}>
-				<ContentsStackProvider />
+				<ContentsStack />
 			</Suspense>
 		</>
 	);
