+++ src/app/(dumper)/@image/page.tsx
@@ -1,10 +1,10 @@
+import { ImageStackSkeleton } from "@/components/stack/image-stack-skeleton";
 import { Separator } from "@/components/ui/separator";
 import { checkSelfAuthOrRedirectToAuth } from "@/features/auth/utils/get-session";
 import { hasDumperPermission } from "@/features/auth/utils/role";
 import { AddFormSkeleton } from "@/features/dump/components/add-form-skeleton";
 import { AddImageProvider } from "@/features/image/components/add-image-provider";
-import { ImageStackProvider } from "@/features/image/components/image-stack-provider";
-import { ImageStackSkeleton } from "@/features/image/components/image-stack-skeleton";
+import { ImageStack } from "@/features/image/components/image-stack";
 import { Suspense } from "react";
 
 export const dynamic = "force-dynamic";
@@ -23,7 +23,7 @@ export default async function Page() {
 			)}
 			<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
 			<Suspense fallback={<ImageStackSkeleton />}>
-				<ImageStackProvider />
+				<ImageStack />
 			</Suspense>
 		</>
 	);
