+++ src/app/(dumper)/@news/page.tsx
@@ -4,7 +4,7 @@ import { checkSelfAuthOrRedirectToAuth } from "@/features/auth/utils/get-session
 import { hasDumperPermission } from "@/features/auth/utils/role";
 import { AddFormSkeleton } from "@/features/dump/components/add-form-skeleton";
 import { AddNewsProvider } from "@/features/news/components/add-news-provider";
-import { NewsStackProvider } from "@/features/news/components/news-stack-provider";
+import { NewsStack } from "@/features/news/components/news-stack";
 import { Suspense } from "react";
 
 export const dynamic = "force-dynamic";
@@ -23,7 +23,7 @@ export default async function Page() {
 			)}
 			<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
 			<Suspense fallback={<CardStackSkeleton />}>
-				<NewsStackProvider />
+				<NewsStack />
 			</Suspense>
 		</>
 	);
