+++ src/components/stack/image-stack-skeleton.tsx
