+++ src/features/news/components/news-stack.tsx
@@ -7,17 +7,19 @@ import { getUserId } from "@/features/auth/utils/get-session";
 import { loggerError } from "@/pino";
 import prisma from "@/prisma";
 
-export async function ContentsStackProvider() {
+export async function NewsStack() {
 	try {
 		const userId = await getUserId();
-		const unexportedContents = (
-			await prisma.contents.findMany({
+
+		const unexportedNews = (
+			await prisma.news.findMany({
 				where: { status: "UNEXPORTED", userId },
 				select: {
 					id: true,
 					title: true,
 					quote: true,
 					url: true,
+					Category: { select: { name: true } },
 				},
 				orderBy: { id: "desc" },
 			})
@@ -27,15 +29,15 @@ export async function ContentsStackProvider() {
 				title: d.title,
 				quote: d.quote,
 				url: d.url,
+				category: d.Category.name,
 			};
 		});
-
-		return <CardStack data={unexportedContents} />;
+		return <CardStack data={unexportedNews} />;
 	} catch (error) {
 		loggerError(
 			ERROR_MESSAGES.UNEXPECTED,
 			{
-				caller: "ContentsStackProvider",
+				caller: "NewsStack",
 				status: 500,
 			},
 			error,
